✨ (#212) 할 일 상세에서 사용되는 댓글 유틸 분리
